refactor: core로 이동한 domain, repository 다시 news 디렉토리로 이동
